requirements.txt
# Core app dependencies
Flask==3.0.3
Werkzeug==3.0.6
Jinja2==3.1.6
itsdangerous==2.2.0
MarkupSafe==2.1.5
blinker==1.8.2
click==8.1.8
bcrypt==4.3.0
python-dotenv==1.0.1

# Testing and development tools
pytest==8.3.5
flake8
black
bandit
djlint
